fullstack-rag-pdf-chat-app/pyproject.toml
[project]
name = "rag-pdf-chat"
version = "1.0.0"
description = "AI-powered RAG chat application for PDF document Q&A with semantic search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "openai>=1.77.0",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.18",
    "jupyter>=1.1.1",
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[tool.setuptools.packages.find]
include = ["aimakerspace*"]

[tool.uv]
dev-dependencies = ["pytest>=7.0.0", "black>=23.0.0", "ruff>=0.1.0"]
